Renaming ForecastUtils to WeatherUtils - Lint Clean up.
